+++ .github/workflows/staticcheck.yml
@@ -11,7 +11,7 @@ jobs:
     name: "Run staticcheck"
     runs-on: ubuntu-latest
     strategy:
-      matrix: { dir: ['.', './server', './server/api' ] }
+      matrix: { dir: ['./server' ] }
     steps:
       - uses: actions/checkout@v3
         with:
