--- go.work
@@ -1,8 +1,0 @@
-go 1.18
-
-use (
-	./api
-	./client
-	./server
-    .
-)
