+++ server/backend/cron/averageRatingComputation.go
@@ -1,7 +1,7 @@
 package cron
 
 import (
-	"github.com/TUM-Dev/Campus-Backend/model"
+	"github.com/TUM-Dev/Campus-Backend/server/model"
 	log "github.com/sirupsen/logrus"
 )
