+++ server/backend/migration/20221210000000.go
@@ -3,7 +3,7 @@ package migration
 import (
 	"database/sql"
 
-	"github.com/TUM-Dev/Campus-Backend/model"
+	"github.com/TUM-Dev/Campus-Backend/server/model"
 	"github.com/go-gormigrate/gormigrate/v2"
 	"github.com/guregu/null"
 	"gorm.io/gorm"
